Code tidy!
