+++ functions/src/admin.ts
@@ -1,0 +1,71 @@
+import { onCall, HttpsError } from "firebase-functions/v2/https";
+import * as logger from "firebase-functions/logger";
+import * as admin from "firebase-admin";
+import { requireAdmin, requireString, getAdminUsers, mapUserRecord } from "./helpers";
+
+/**
+ * Grants admin claim to a user
+ */
+export const grantAdmin = onCall(async (request) => {
+  requireAdmin(request);
+  const uid = requireString(request.data.uid, "uid");
+
+  try {
+    await admin.auth().setCustomUserClaims(uid, { admin: true });
+    logger.info(`Admin claim set for uid=${uid} by caller=${request.auth!.uid}`);
+    return { success: true, message: `Admin claim set for uid=${uid}` };
+  } catch (e: any) {
+    logger.error("Error setting custom claim:", e);
+    throw new HttpsError("internal", e?.message ?? "Error setting custom claim");
+  }
+});
+
+/**
+ * Revokes admin claim from a user, ensuring at least one admin remains
+ */
+export const revokeAdmin = onCall(async (request) => {
+  requireAdmin(request);
+  const uid = requireString(request.data.uid, "uid");
+
+  try {
+    const adminUsers = await getAdminUsers();
+
+    if (adminUsers.length <= 1) {
+      throw new HttpsError("failed-precondition", "Cannot remove the last admin. At least one admin must remain.");
+    }
+
+    const targetUser = await admin.auth().getUser(uid);
+    const targetUserClaims = targetUser.customClaims || {};
+    if (targetUserClaims.admin !== true) {
+      throw new HttpsError("failed-precondition", "User is not an admin");
+    }
+
+    await admin.auth().setCustomUserClaims(uid, { admin: false });
+    logger.info(`Admin claim removed for uid=${uid} by caller=${request.auth!.uid}`);
+    return { success: true, message: `Admin claim removed for uid=${uid}` };
+  } catch (e: any) {
+    if (e instanceof HttpsError) {
+      throw e;
+    }
+    logger.error("Error removing custom claim:", e);
+    throw new HttpsError("internal", e?.message ?? "Error removing custom claim");
+  }
+});
+
+/**
+ * Lists all users with admin claim
+ */
+export const listAdminUsers = onCall(async (request) => {
+  requireAdmin(request);
+
+  try {
+    const adminUsers = await getAdminUsers();
+    const mappedUsers = adminUsers.map(mapUserRecord);
+    logger.info(`Found ${mappedUsers.length} admin users for caller ${request.auth!.uid}`);
+    return { users: mappedUsers };
+  } catch (e: any) {
+    logger.error("Error listing admin users:", e);
+    throw new HttpsError("internal", e?.message ?? "Error listing admin users");
+  }
+});
+
