+++ functions/src/helpers.ts
@@ -1,0 +1,62 @@
+import { HttpsError } from "firebase-functions/v2/https";
+import type { CallableRequest } from "firebase-functions/v2/https";
+import * as admin from "firebase-admin";
+
+/**
+ * Ensures the request is authenticated
+ */
+export function requireAuth(request: CallableRequest): void {
+  if (!request.auth) {
+    throw new HttpsError("unauthenticated", "Sign in required");
+  }
+}
+
+/**
+ * Ensures the request is authenticated and the user is an admin
+ */
+export function requireAdmin(request: CallableRequest): void {
+  requireAuth(request);
+  if (request.auth!.token.admin !== true) {
+    throw new HttpsError("permission-denied", "Admins only");
+  }
+}
+
+/**
+ * Validates that a value is a non-empty string
+ */
+export function requireString(value: any, fieldName: string): string {
+  if (!value || typeof value !== "string" || value.trim().length === 0) {
+    throw new HttpsError("invalid-argument", `${fieldName} is required and must be a non-empty string`);
+  }
+  return value.trim();
+}
+
+/**
+ * Maps a Firebase Auth user record to a simplified user object
+ */
+export function mapUserRecord(userRecord: admin.auth.UserRecord) {
+  return {
+    uid: userRecord.uid,
+    email: userRecord.email || "",
+    displayName: userRecord.displayName || "",
+    emailVerified: userRecord.emailVerified,
+    disabled: userRecord.disabled,
+    metadata: {
+      creationTime: userRecord.metadata.creationTime,
+      lastSignInTime: userRecord.metadata.lastSignInTime,
+    },
+    customClaims: userRecord.customClaims || {},
+  };
+}
+
+/**
+ * Gets all admin users from Firebase Auth
+ */
+export async function getAdminUsers(): Promise<admin.auth.UserRecord[]> {
+  const listUsersResult = await admin.auth().listUsers();
+  return listUsersResult.users.filter((userRecord) => {
+    const customClaims = userRecord.customClaims || {};
+    return customClaims.admin === true;
+  });
+}
+
